V1.1.2 Removed tags for JSON omitempty. The field of the signInput structure has been changed. Implemented a request to YTimes to receive points when entering the system (signIn) with filling in the response in the server database.

+++ schema/000001_init.up.sql
@@ -7,7 +7,7 @@ CREATE TABLE users
     surname       varchar(50),
     email         varchar(50),
     birthday      date,
-    value         int          default 0,
+    value         float        default 0,
     message_key   varchar(255) unique
 );
